feat: move around, new welcome with bg only, and new start destination
