+++ app/src/main/kotlin/com/wire/android/ui/authentication/welcome/WelcomeScreen.kt
@@ -20,60 +20,28 @@
 
 package com.wire.android.ui.authentication.welcome
 
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.interaction.MutableInteractionSource
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.ExperimentalComposeUiApi
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.platform.testTag
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.res.vectorResource
-import androidx.compose.ui.semantics.semantics
-import androidx.compose.ui.semantics.testTagsAsResourceId
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.text.style.TextDecoration
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.ramcosta.composedestinations.annotation.RootNavGraph
-import com.wire.android.R
-import com.wire.android.config.LocalCustomUiConfigurationProvider
 import com.wire.android.navigation.NavigationCommand
 import com.wire.android.navigation.Navigator
 import com.wire.android.navigation.WireDestination
 import com.wire.android.navigation.style.PopUpNavigationAnimation
-import com.wire.android.ui.authentication.ServerTitle
-import com.wire.android.ui.authentication.login.NewLoginContainer
-import com.wire.android.ui.common.button.WirePrimaryButton
-import com.wire.android.ui.common.button.WireSecondaryButton
+import com.wire.android.ui.MainBackgroundComponent
 import com.wire.android.ui.common.dialogs.FeatureDisabledWithProxyDialogContent
 import com.wire.android.ui.common.dialogs.FeatureDisabledWithProxyDialogState
 import com.wire.android.ui.common.dialogs.MaxAccountsReachedDialog
 import com.wire.android.ui.common.dialogs.MaxAccountsReachedDialogState
-import com.wire.android.ui.common.dimensions
 import com.wire.android.ui.common.visbility.rememberVisibilityState
-import com.wire.android.ui.destinations.CreatePersonalAccountOverviewScreenDestination
-import com.wire.android.ui.destinations.CreateTeamAccountOverviewScreenDestination
 import com.wire.android.ui.destinations.LoginScreenDestination
 import com.wire.android.ui.theme.WireTheme
-import com.wire.android.ui.theme.wireDimensions
-import com.wire.android.ui.theme.wireTypography
 import com.wire.android.util.CustomTabsHelper
 import com.wire.android.util.ui.PreviewMultipleThemes
 import com.wire.kalium.logic.configuration.server.ServerConfig
+import kotlinx.coroutines.delay
 
 @RootNavGraph(start = true)
 @WireDestination(
@@ -85,9 +53,7 @@ fun WelcomeScreen(
     viewModel: WelcomeViewModel = hiltViewModel()
 ) {
     WelcomeContent(
-        viewModel.state.isThereActiveSession,
-        viewModel.state.maxAccountsReached,
-        viewModel.state.links,
+        viewModel.state,
         navigator::navigateBack,
         navigator::navigate
     )
@@ -95,161 +61,35 @@ fun WelcomeScreen(
 
 @Composable
 private fun WelcomeContent(
-    isThereActiveSession: Boolean,
-    maxAccountsReached: Boolean,
-    state: ServerConfig.Links,
+    state: WelcomeScreenState,
     navigateBack: () -> Unit,
     navigate: (NavigationCommand) -> Unit
 ) {
-    NewLoginContainer(
-        title = stringResource(R.string.enterprise_login_welcome),
-        canNavigateBack = isThereActiveSession,
-        onNavigateBack = navigateBack
-    ) {
-        NewWelcomeExperienceContent(
-            navigateBack = navigateBack,
-            maxAccountsReached = maxAccountsReached,
-            state = state,
-            navigate = navigate
-        )
-    }
-}
-
-@OptIn(ExperimentalComposeUiApi::class)
-@Composable
-private fun NewWelcomeExperienceContent(
-    navigateBack: () -> Unit,
-    maxAccountsReached: Boolean,
-    state: ServerConfig.Links,
-    navigate: (NavigationCommand) -> Unit,
-) {
+    MainBackgroundComponent()
     val enterpriseDisabledWithProxyDialogState = rememberVisibilityState<FeatureDisabledWithProxyDialogState>()
     val createPersonalAccountDisabledWithProxyDialogState = rememberVisibilityState<FeatureDisabledWithProxyDialogState>()
     val context = LocalContext.current
-    Column(
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.SpaceBetween,
-        modifier = Modifier
-    ) {
-        val maxAccountsReachedDialogState = rememberVisibilityState<MaxAccountsReachedDialogState>()
-        MaxAccountsReachedDialog(dialogState = maxAccountsReachedDialogState) { navigateBack() }
-        if (maxAccountsReached) {
-            maxAccountsReachedDialogState.show(maxAccountsReachedDialogState.savedState ?: MaxAccountsReachedDialogState)
+    val maxAccountsReachedDialogState = rememberVisibilityState<MaxAccountsReachedDialogState>()
+    MaxAccountsReachedDialog(dialogState = maxAccountsReachedDialogState) { navigateBack() }
+    if (state.maxAccountsReached) {
+        maxAccountsReachedDialogState.show(maxAccountsReachedDialogState.savedState ?: MaxAccountsReachedDialogState)
+    }
+    FeatureDisabledWithProxyDialogContent(
+        dialogState = enterpriseDisabledWithProxyDialogState,
+        onActionButtonClicked = {
+            CustomTabsHelper.launchUrl(context, state.links.teams)
         }
+    )
+    FeatureDisabledWithProxyDialogContent(dialogState = createPersonalAccountDisabledWithProxyDialogState)
 
-        Icon(
-            imageVector = ImageVector.vectorResource(id = R.drawable.ic_wire_logo),
-            tint = MaterialTheme.colorScheme.onBackground,
-            contentDescription = stringResource(id = R.string.content_description_welcome_wire_logo),
-            modifier = Modifier.size(dimensions().spacing120x)
-        )
-
-        if (state.isOnPremises) {
-            ServerTitle(serverLinks = state, modifier = Modifier.padding(top = dimensions().spacing16x))
-        }
-
-        Column(
-            modifier = Modifier
-                .padding(
-                    vertical = MaterialTheme.wireDimensions.welcomeVerticalSpacing,
-                    horizontal = MaterialTheme.wireDimensions.welcomeButtonHorizontalPadding
-                )
-                .semantics {
-                    testTagsAsResourceId = true
-                }
-        ) {
-            LoginButton(onClick = { navigate(NavigationCommand(LoginScreenDestination())) })
-            FeatureDisabledWithProxyDialogContent(
-                dialogState = enterpriseDisabledWithProxyDialogState,
-                onActionButtonClicked = {
-                    CustomTabsHelper.launchUrl(context, state.teams)
-                }
-            )
-            FeatureDisabledWithProxyDialogContent(dialogState = createPersonalAccountDisabledWithProxyDialogState)
-
-            if (LocalCustomUiConfigurationProvider.current.isAccountCreationAllowed) {
-                CreateEnterpriseAccountButton {
-                    if (state.isProxyEnabled()) {
-                        enterpriseDisabledWithProxyDialogState.show(
-                            enterpriseDisabledWithProxyDialogState.savedState ?: FeatureDisabledWithProxyDialogState(
-                                R.string.create_team_not_supported_dialog_description,
-                                state.teams
-                            )
-                        )
-                    } else {
-                        navigate(NavigationCommand(CreateTeamAccountOverviewScreenDestination))
-                    }
-                }
+    LaunchedEffect(Unit) {
+        if (state.maxAccountsReached.not()) {
+            delay(1_000) // small delay to resolve
+            when (state.startLoginDestination) {
+                StartLoginDestination.Default -> navigate(NavigationCommand(LoginScreenDestination()))
+                StartLoginDestination.CustomBackend -> navigate(NavigationCommand(LoginScreenDestination())) // todo pass parameter from deeplink processor
             }
         }
-
-        if (LocalCustomUiConfigurationProvider.current.isAccountCreationAllowed) {
-            WelcomeFooter(
-                modifier = Modifier.padding(horizontal = MaterialTheme.wireDimensions.welcomeTextHorizontalPadding),
-                onPrivateAccountClick = {
-                    if (state.isProxyEnabled()) {
-                        createPersonalAccountDisabledWithProxyDialogState.show(
-                            createPersonalAccountDisabledWithProxyDialogState.savedState ?: FeatureDisabledWithProxyDialogState(
-                                R.string.create_personal_account_not_supported_dialog_description
-                            )
-                        )
-                    } else {
-                        navigate(NavigationCommand(CreatePersonalAccountOverviewScreenDestination))
-                    }
-                }
-            )
-        }
-    }
-}
-
-@Composable
-private fun LoginButton(onClick: () -> Unit) {
-    WirePrimaryButton(
-        onClick = onClick,
-        text = stringResource(R.string.label_login),
-        modifier = Modifier
-            .padding(bottom = MaterialTheme.wireDimensions.welcomeButtonVerticalPadding)
-            .testTag("loginButton")
-    )
-}
-
-@Composable
-private fun CreateEnterpriseAccountButton(onClick: () -> Unit) {
-    WireSecondaryButton(
-        onClick = onClick,
-        text = stringResource(R.string.welcome_button_create_team),
-        modifier = Modifier.padding(bottom = MaterialTheme.wireDimensions.welcomeButtonVerticalPadding)
-    )
-}
-
-@Composable
-private fun WelcomeFooter(onPrivateAccountClick: () -> Unit, modifier: Modifier = Modifier) {
-    Column(modifier = modifier) {
-        Text(
-            text = stringResource(R.string.welcome_footer_text),
-            style = MaterialTheme.wireTypography.body02,
-            textAlign = TextAlign.Center,
-            modifier = Modifier.fillMaxWidth()
-        )
-
-        Text(
-            text = stringResource(R.string.welcome_button_create_personal_account),
-            style = MaterialTheme.wireTypography.body02.copy(
-                textDecoration = TextDecoration.Underline,
-                color = MaterialTheme.colorScheme.primary
-            ),
-            textAlign = TextAlign.Center,
-            modifier = Modifier
-                .fillMaxWidth()
-                .clickable(
-                    interactionSource = remember { MutableInteractionSource() },
-                    indication = null,
-                    onClick = onPrivateAccountClick,
-                    onClickLabel = stringResource(R.string.content_description_open_link_label)
-                )
-        )
-
-        Spacer(modifier = Modifier.height(MaterialTheme.wireDimensions.welcomeVerticalPadding))
     }
 }
 
@@ -259,9 +99,7 @@ private fun WelcomeFooter(onPrivateAccountClick: () -> Unit, modifier: Modifier 
 fun PreviewWelcomeScreen() {
     WireTheme {
         WelcomeContent(
-            isThereActiveSession = false,
-            maxAccountsReached = false,
-            state = ServerConfig.DEFAULT,
+            state = WelcomeScreenState(ServerConfig.DEFAULT),
             navigateBack = {},
             navigate = {})
     }
